Refactored graph manager
