+++ UI/src/mainwindow/mainwindow.cpp
@@ -1,4 +1,4 @@
-#include "mainwindow.h"
+#include "mainwindow/mainwindow.h"
 
 #include <iostream>
 
@@ -19,6 +19,8 @@ MainWindow::MainWindow(QWidget *parent)
   InitializeEngine();
   PauseSimulation();
   RunSimulation();
+
+  graph_manager_ = new GraphManager(this, engine_);
 
   // Assuming you have a QComboBox named 'yourComboBoxName' in your UI
   QComboBox *comboBox = ui_->graphMenu;
@@ -71,7 +73,7 @@ MainWindow::MainWindow(QWidget *parent)
           &MainWindow::ShowConfigScreen);
   connect(ui_->restartButton, &QPushButton::clicked, this,
           &MainWindow::RestartSimulation);
-  connect(ui_->graphMenu, SIGNAL(currentIndexChanged(int)), this, SLOT(handleDropdownSelection(int)));
+  connect(ui_->graphMenu, SIGNAL(currentIndexChanged(int)), graph_manager_, SLOT(handleDropdownSelection(int)));
   updateTimer = new QTimer(this);
   connect(updateTimer, SIGNAL(timeout()), this, SLOT(recordCreatureCount()));
   updateTimer->start(1000);  // Set the interval to 1000 milliseconds (1 second)
@@ -281,92 +283,7 @@ void MainWindow::RestartSimulation() {
 }
 
 
-void MainWindow::DrawCreaturesOverTimeGraph() {
-  if (engine_->GetSimulation()) {
-    // Get the creature count over time from the simulation data
-    std::vector<int> creatureCountOverTime = engine_->GetSimulation()->GetSimulationData()->GetCreatureCountOverTime();
 
-    // Check if there's any data to display
-    if (creatureCountOverTime.empty()) {
-        qDebug() << "No data to display.";
-        return;
-    }
-
-    // Create a new line series
-    QLineSeries *series = new QLineSeries();
-
-    // Add data points to the series
-    for (size_t i = 0; i < creatureCountOverTime.size(); ++i) {
-        series->append(i, creatureCountOverTime[i]);
-    }
-
-    // Create a chart and add the series to it
-    QChart *chart = new QChart();
-    chart->addSeries(series);
-    chart->createDefaultAxes();
-    chart->axes(Qt::Horizontal).first()->setTitleText("Time Elapsed");
-    chart->axes(Qt::Vertical).first()->setTitleText("Number of Creatures in the Simulation");
-
-    // Create a chart view with the chart
-    QChartView *chartView = new QChartView(chart);
-    chartView->setRenderHint(QPainter::Antialiasing);
-
-    // Create a dialog to display the graph
-    QDialog *dialog = new QDialog(this);
-    dialog->setWindowTitle("Creature Count Over Time");
-    QVBoxLayout *layout = new QVBoxLayout(dialog);
-    layout->addWidget(chartView);
-    dialog->setLayout(layout);
-    dialog->resize(800, 600);
-
-    // Show the dialog modally and connect it to delete later
-    connect(dialog, &QDialog::finished, dialog, &QObject::deleteLater);
-    dialog->exec();
-  }
-}
-void MainWindow::DrawCreaturesSizeOverTimeGraph() {
-  if (engine_->GetSimulation()) {
-    // Get the creature count over time from the simulation data
-    std::vector<int> creatureCountOverTime = engine_->GetSimulation()->GetSimulationData()->GetCreatureSizeOverTime();
-
-    // Check if there's any data to display
-    if (creatureCountOverTime.empty()) {
-        qDebug() << "No data to display.";
-        return;
-    }
-
-    // Create a new line series
-    QLineSeries *series = new QLineSeries();
-
-    // Add data points to the series
-    for (size_t i = 0; i < creatureCountOverTime.size(); ++i) {
-        series->append(i, creatureCountOverTime[i]);
-    }
-
-    // Create a chart and add the series to it
-    QChart *chart = new QChart();
-    chart->addSeries(series);
-    chart->createDefaultAxes();
-    chart->axes(Qt::Horizontal).first()->setTitleText("Time Elapsed");
-    chart->axes(Qt::Vertical).first()->setTitleText("Average Size of Creatures in the Simulation");
-
-    // Create a chart view with the chart
-    QChartView *chartView = new QChartView(chart);
-    chartView->setRenderHint(QPainter::Antialiasing);
-
-    // Create a dialog to display the graph
-    QDialog *dialog = new QDialog(this);
-    dialog->setWindowTitle("Average Creature Size Over Time");
-    QVBoxLayout *layout = new QVBoxLayout(dialog);
-    layout->addWidget(chartView);
-    dialog->setLayout(layout);
-    dialog->resize(800, 600);
-
-    // Show the dialog modally and connect it to delete later
-    connect(dialog, &QDialog::finished, dialog, &QObject::deleteLater);
-    dialog->exec();
-  }
-}
 
 void MainWindow::keyPressEvent(QKeyEvent *event) {
     if (event->key() == Qt::Key_Escape) {
@@ -380,17 +297,3 @@ void MainWindow::keyPressEvent(QKeyEvent *event) {
     }
 }
 
-void MainWindow::handleDropdownSelection(int index) {
-    // Here, you can call the function corresponding to the selected item
-    qDebug() << "Dropdown selection changed to index:" << index;
-
-    // Check if the selected index is 1, then call DrawCreaturesOverTimeGraph
-    if (index == 1) {
-        qDebug() << "Calling DrawCreaturesOverTimeGraph";
-        DrawCreaturesOverTimeGraph();
-    }
-    if (index == 2) {
-        qDebug() << "Calling DrawCreaturesOverTimeGraph";
-        DrawCreaturesSizeOverTimeGraph();
-    }
-}
